add: implement trace support
